remove danger from other files
